src/tools-TEST/* fixed bug in CI tests
