+++ src/tools-TEST/tools-TEST.pro
@@ -128,7 +128,6 @@ HEADERS += NGSDAddVariantsSomatic_Test.h \
     SplicingToBed_Test.h \
     GraphStringDb_Test.h \
     GenePrioritization_Test.h \
-    CfDnaQC_Test.h \
     VcfAnnotateFromBigWig_Test.h \
     BedLiftOver_Test.h \
     BedpeSort_Test.h \
@@ -136,6 +135,7 @@ HEADERS += NGSDAddVariantsSomatic_Test.h \
     BedpeAnnotateCounts_Test.h \
     BedpeAnnotateBreakpointDensity.h \
     VcfAnnotateConsequence_Test.h \
+    CfnaQC_Test.h \
     RnaQC_Test.h
 
 SOURCES += \
